requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.1.0=pyhd8ed1ab_0
aiohttp=3.8.1=py39hb9d737c_1
aiosignal=1.2.0=pyhd8ed1ab_0
alsa-lib=1.2.6.1=h7f98852_0
aom=3.3.0=h27087fc_1
async-timeout=4.0.2=pyhd8ed1ab_0
attr=2.5.1=h166bdaf_0
attrs=21.4.0=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.4=py_1
bottleneck=1.3.4=py39hce1f21e_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py39hb9d737c_1004
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.0.0=pyhd8ed1ab_0
cairo=1.16.0=ha61ee94_1011
certifi=2022.6.15=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.3=py39hf3d152e_0
cryptography=37.0.2=py39hd97740a_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd8ed1ab_0
daal4py=2021.5.0=py39h78b71dc_0
dal=2021.5.1=h06a4308_803
dbus=1.13.6=he372182_0
easydict=1.9=py_0
expat=2.4.8=h27087fc_0
ffmpeg=4.4.2=habc3f16_0
fftw=3.3.8=nompi_hfc0cae8_1114
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=h8e229c2_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.33.3=py39hb9d737c_0
freeglut=3.2.2=h9c3ff4c_1
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=py39hb9d737c_1
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h36c2ea0_2
glib=2.72.1=h6239696_0
glib-tools=2.72.1=h6239696_0
gmp=6.2.1=h58526e2_0
gnutls=3.7.6=hf3e180e_5
google-auth=2.8.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.42.0=py39hce63b2e_0
gst-plugins-base=1.20.3=hf6a322e_0
gstreamer=1.20.3=hd4edc92_0
harfbuzz=4.4.1=hf9f4e7c_0
hdf5=1.12.1=h69dfa17_1
icu=70.1=h27087fc_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.4=py39hf3d152e_0
intel-openmp=2021.4.0=h06a4308_3561
jack=1.9.18=h8c3723f_1002
jasper=2.0.33=ha77e612_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h166bdaf_1
json-c=0.16=hc379101_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.2=py39h295c915_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcap=2.64=ha37c62d_0
libcblas=3.9.0=12_linux64_mkl
libclang=14.0.6=default_h2e3cab8_0
libclang13=14.0.6=default_h3a83d3e_0
libcups=2.3.3=hf5a7f15_1
libcurl=7.83.1=h7bff187_0
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.12=h166bdaf_0
libdrm=2.4.112=h166bdaf_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h7f98852_5
libflac=1.3.4=h27087fc_0
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglib=2.72.1=h2d90d5f_0
libglu=9.0.0=he1b5a44_1001
libiconv=1.16=h516909a_0
libidn2=2.3.3=h166bdaf_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm14=14.0.6=he0ac6c6_0
libnghttp2=1.47.0=h727a467_0
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopencv=4.6.0=py39hd011c1b_0
libopus=1.3.1=h7f98852_1
libpciaccess=0.16=h516909a_0
libpng=1.6.37=h21135ba_2
libpq=14.4=hd77ab85_0
libprotobuf=3.20.1=h6239696_0
libsndfile=1.0.31=h9c3ff4c_1
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=12.1.0=ha89aaad_16
libtasn1=4.18.0=h166bdaf_1
libtiff=4.4.0=hc85c160_1
libtool=2.4.6=h9c3ff4c_1008
libunistring=0.9.10=h7f98852_0
libuuid=2.32.1=h7f98852_1000
libuv=1.43.0=h7f98852_0
libva=2.15.0=h166bdaf_0
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.11.0=h9c3ff4c_3
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.14=h22db469_3
libzlib=1.2.12=h166bdaf_2
llvm-openmp=14.0.4=he0ac6c6_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.7=pyhd8ed1ab_0
matplotlib=3.5.1=py39h06a4308_1
matplotlib-base=3.5.1=py39ha18d171_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7e14d7c_0
mkl_fft=1.3.1=py39h0c7bc48_1
mkl_random=1.2.2=py39hde0f152_0
mpi=1.0=mpich
mpich=3.3.2=hc856adb_0
multidict=6.0.2=py39hb9d737c_1
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.30=haf5c9bc_0
mysql-libs=8.0.30=h28c427c_0
ncurses=6.3=h27087fc_1
nettle=3.8.1=hc379101_1
nspr=4.32=h9c3ff4c_1
nss=3.78=h2350873_0
numexpr=2.8.1=py39h6abb31d_0
numpy=1.22.3=py39he7a7128_0
numpy-base=1.22.3=py39hf524024_0
oauthlib=3.2.0=pyhd8ed1ab_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.6.0=py39hf3d152e_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1q=h7f8727e_0
p11-kit=0.24.1=hc5aa10d_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py39h295c915_0
pcre=8.45=h9c3ff4c_0
pillow=9.2.0=py39hae2aec6_0
pip=22.1.2=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
ply=3.11=py_1
portaudio=19.6.0=h57a0ea0_5
protobuf=4.21.1=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pulseaudio=14.0=h583eb01_5
py-opencv=4.6.0=py39hef51801_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.4.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py39h18e9c17_0
pyqt5-sip=12.11.0=py39h5a03fae_0
pysocks=1.7.1=py39hf3d152e_5
python=3.9.13=h9a8a25e_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.10.1=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py39h7f8727e_1
qt-main=5.15.4=ha5833f6_2
readline=8.1.2=h7f8727e_1
requests=2.28.0=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.0.2=py39h51133e4_1
scikit-learn-intelex=2021.5.0=py39h06a4308_0
scipy=1.7.3=py39hc147768_0
setuptools=59.5.0=py39hf3d152e_0
sip=6.6.2=py39h5a03fae_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.39.2=h4ff8645_0
svt-av1=1.1.0=h27087fc_1
tbb=2021.5.0=hd09550d_0
tensorboard=2.8.0=py39h06a4308_0
tensorboard-data-server=0.6.0=py39hd97740a_2
tensorboard-logger=0.1.0=pypi_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
terminaltables=3.1.10=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h27826a3_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=0.10.1=py39_cu113
torchvision=0.11.2=py39_cu113
tornado=6.1=py39hb9d737c_3
tqdm=4.64.0=py39h06a4308_0
typing-extensions=4.2.0=hd8ed1ab_1
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=h191b570_0
unicodedata2=14.0.0=py39hb9d737c_1
urllib3=1.26.9=pyhd8ed1ab_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.37.1=pyhd8ed1ab_0
x264=1!161.3030=h7f98852_1
x265=3.5=h924138e_3
xcb-util=0.4.0=h166bdaf_0
xcb-util-image=0.4.0=h166bdaf_0
xcb-util-keysyms=0.4.0=h166bdaf_0
xcb-util-renderutil=0.3.9=h166bdaf_0
xcb-util-wm=0.4.1=h166bdaf_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py39hb9d737c_2
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h166bdaf_2
zstd=1.5.2=h8a70e8d_2
